LPS-61466 split SearchEngineHelperUtil into indexSearcherHelper, indexWriterHelper
