+++ portal-service/src/com/liferay/portal/kernel/search/IndexSearcherHelper.java
@@ -1,0 +1,43 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.kernel.search;
+
+import java.util.List;
+import java.util.Map;
+
+/**
+ * @author Michael C. Han
+ */
+public interface IndexSearcherHelper {
+
+	public String getQueryString(SearchContext searchContext, Query query);
+
+	public Hits search(SearchContext searchContext, Query query)
+		throws SearchException;
+
+	public long searchCount(SearchContext searchContext, Query query)
+		throws SearchException;
+
+	public String spellCheckKeywords(SearchContext searchContext)
+		throws SearchException;
+
+	public Map<String, List<String>> spellCheckKeywords(
+			SearchContext searchContext, int max)
+		throws SearchException;
+
+	public String[] suggestKeywordQueries(SearchContext searchContext, int max)
+		throws SearchException;
+
+}
